dom/media/mediasource/test/test_SeekableBeforeAndAfterEndOfStream_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute after end of stream</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/mp4");

  const arrayBuffer = await fetchWithXHR("bipbop/bipbop2s.mp4");
  info("- append buffer -");
  sb.appendBuffer(new Uint8Array(arrayBuffer));

  info("- wait for metadata -");
  await once(v, "loadedmetadata");

  info("- wait for updateend -");
  await once(sb, "updateend");

  info("- check seekable -");
  const target = 1.3;
  ok(v.seekable.length, "Resource is seekable");
  is(v.seekable.start(0), 0, "Seekable's start point is correct");
  is(v.seekable.end(0), ms.duration, "Seekable's end point is correct");
  ok(v.seekable.length &&
     target >= v.seekable.start(0) &&
     target < v.seekable.end(0), "Target is within seekable range");

  info("- call end of stream -");
  ms.endOfStream();
  await once(ms, "sourceended");

  info("- check seekable -");
  ok(v.seekable.length, "Resource is seekable");
  is(v.seekable.start(0), 0, "Seekable's start point is correct");
  is(v.seekable.end(0), ms.duration, "Seekable's end point is correct");
  ok(v.seekable.length &&
     target >= v.seekable.start(0) &&
     target < v.seekable.end(0), "Target is within seekable range");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
